+++ docs/.vitepress/config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from "vitepress";
 import llmstxt from "vitepress-plugin-llms";
+// import { createRewrites } from "../../packages/config";
 // 本地 Teek 主题包引用（与 Teek 在线主题包引用 二选一）
 import { defineTeekConfig } from "../../packages/config";
 import { version } from "../../packages/teek/version";
@@ -44,6 +45,7 @@ const teekConfig = defineTeekConfig({
   vitePlugins: {
     sidebarOption: {
       initItems: false,
+      // resolveRule: "rewrites",
     },
   },
   markdown: {
@@ -134,6 +136,7 @@ export default defineConfig({
       return [...items, ...permalinkItemBak];
     },
   },
+  // rewrites: createRewrites(),
   themeConfig: {
     // https://vitepress.dev/reference/default-theme-config
     logo: "/teek-logo-mini.svg",
@@ -161,9 +164,9 @@ export default defineConfig({
       {
         text: "资源",
         items: [
-          { text: "案例", link: "/case" },
-          { text: "常见问题", link: "/theme/qa" },
-          { text: "功能拓展", link: "/expand/intro" },
+          { text: "案例", link: "/resources/case" },
+          { text: "常见问题", link: "/resources/qa" },
+          { text: "功能拓展", link: "/resources/expand/intro" },
         ],
       },
       {