+++ docs/.vitepress/theme/components/TeekLayoutProvider.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts" name="TeekLayoutProvider">
 import type { TeekConfig } from "vitepress-theme-teek";
-import Teek, { teekConfigContext, clockIcon } from "vitepress-theme-teek";
+import Teek, { teekConfigContext, clockIcon, TkPageCard } from "vitepress-theme-teek";
 import { useData } from "vitepress";
 import { watch, nextTick, ref, provide } from "vue";
 import { teekDocConfig } from "../config/teekConfig";
@@ -47,6 +47,8 @@ const handleConfigSwitch = (config: TeekConfig, style: string) => {
 
   watchRuntimeAndRibbon(frontmatter.value.layout, style);
 };
+
+const list = ["name1", "name2", "name3", "name4"];
 </script>
 
 <template>