+++ packages/components/theme/CataloguePage/src/CatalogueItem.vue
@@ -14,7 +14,7 @@ defineProps<{ item: CatalogueItem; index: number | string }>();
 
 <template>
   <li :class="item.children ? nsSub.b() : nsItem.b()">
-    <a v-if="!item.children" :href="item.link && withBase(item.link)" :aria-label="`${index}. ${item.title}`">
+    <a v-if="!item.children" :href="item.url && withBase(item.url)" :aria-label="`${index}. ${item.title}`">
       {{ index }}.
       <span v-html="item.title" />
       <TkTitleTag
@@ -26,7 +26,7 @@ defineProps<{ item: CatalogueItem; index: number | string }>();
       />
     </a>
 
-    <template v-else-if="item.children.length > 0">
+    <template v-else-if="item.children.length">
       <div :id="item.title" :class="nsSub.e('title')" role="group" :aria-labelledby="`${item.title}-label`">
         <a :href="`#${item.title}`" class="anchor" :aria-label="item.title">#</a>
         <span :id="`${item.title}-label`">{{ `${index}. ${item.title}` }}</span>