+++ plugins/vitepress-plugin-sidebar-resolve/src/index.ts
@@ -1,12 +1,13 @@
 import type { Plugin, ViteDevServer } from "vite";
 import type { SidebarOption } from "./types";
-import createSidebar from "./helper";
-import { join } from "node:path";
 import { DefaultTheme } from "vitepress";
+import { join } from "node:path";
+import createFilePathSidebar from "./filePathToSidebar";
+import createRewritesSidebar from "./rewritesToSidebar";
 import logger from "./log";
 
 export * from "./types";
-export * from "./util";
+export * from "./utils";
 
 export default function VitePluginVitePressSidebarResolve(option: SidebarOption = {}): Plugin & { name: string } {
   let isExecute = false;
@@ -40,47 +41,93 @@ export default function VitePluginVitePressSidebarResolve(option: SidebarOption 
       const {
         site: { themeConfig = {}, locales = {} },
         srcDir,
+        rewrites: rewritesObj,
       } = config.vitepress;
 
-      const { path, ignoreList, localeRootDir } = option;
+      const { path, ignoreList, localeRootDir, type = "object", resolveRule = "filePath" } = option;
       const baseDir = path ? join(srcDir, path) : srcDir;
+      const rewrites = rewritesObj.map || {};
+      const rewritesLength = Object.keys(rewrites).length;
+
+      // 如果 指定 rewrites 规则，但是 rewrites 不存在，则走 filePath 逻辑
+      const isFilePathRule = resolveRule === "filePath" || (resolveRule === "rewrites" && !rewritesLength);
+
+      const isRewritesRule = resolveRule === "rewrites" && rewritesLength;
 
       // 国际化多语言 key 数组
       const localesKeys = Object.keys(locales).filter(key => key !== "root");
 
-      // 如果不是多语言，直接自动生成结构化侧边栏
-      if (!localesKeys.length) return setSideBar(themeConfig, createSidebar({ ...option, path: baseDir }));
+      //  filePath 规则
+      if (isFilePathRule) {
+        // 如果不是多语言，直接自动生成结构化侧边栏
+        if (!localesKeys.length) {
+          return setSideBar(themeConfig, createFilePathSidebar({ ...option, path: baseDir }), type);
+        }
 
-      // 国际化处理，针对每个语言的目录进行单独的扫描（除了 root）
-      localesKeys.forEach(localesKey => {
-        const sidebar: DefaultTheme.SidebarMulti = createSidebar(
-          { ...option, path: `${baseDir}/${localesKey}` },
-          localesKey
-        );
-        setSideBar(locales[localesKey].themeConfig, sidebar);
-      });
+        // 国际化处理，针对每个语言的目录进行单独的扫描（除了 root）
+        localesKeys.forEach(localesKey => {
+          const sidebar = createFilePathSidebar({ ...option, path: `${baseDir}/${localesKey}` }, localesKey);
+          setSideBar(locales[localesKey].themeConfig, sidebar, type);
+        });
 
-      // 如果指定了 root 的目录，则扫描指定目录
-      const rootDir = localeRootDir ? `/${localeRootDir}` : "";
-      // 对 root 根目录的 sidebar 进行单独的扫描，且不扫描其他语言目录
-      const rootSideBar = createSidebar({
-        ...option,
-        path: `${baseDir}${rootDir}`,
-        ignoreList: [...(ignoreList || []), ...localesKeys],
-      });
-      setSideBar(locales["root"].themeConfig, rootSideBar);
+        // 如果指定了 root 的目录，则扫描指定目录
+        const rootDir = localeRootDir ? `/${localeRootDir}` : "";
+        // 对 root 根目录的 sidebar 进行单独的扫描，且不扫描其他语言目录
+        const rootSideBar = createFilePathSidebar({
+          ...option,
+          path: `${baseDir}${rootDir}`,
+          ignoreList: [...(ignoreList || []), ...localesKeys],
+        });
+        setSideBar(locales["root"].themeConfig, rootSideBar, type);
+      }
+      // rewrites 规则
+      else if (isRewritesRule) {
+        // 如果不是多语言，直接自动生成结构化侧边栏
+        if (!localesKeys.length) {
+          return setSideBar(themeConfig, createRewritesSidebar(rewrites, { ...option, path: baseDir }), type);
+        }
+
+        // 国际化处理，针对每个语言的目录进行单独的扫描（除了 root）
+        localesKeys.forEach(localesKey => {
+          const sidebar = createRewritesSidebar(rewrites, { ...option, path: `${baseDir}/${localesKey}` }, localesKey);
+          setSideBar(locales[localesKey].themeConfig, sidebar, type);
+        });
+
+        // 如果指定了 root 的目录，则扫描指定目录
+        const rootDir = localeRootDir ? `/${localeRootDir}` : "";
+        // 对 root 根目录的 sidebar 进行单独的扫描，且不扫描其他语言目录
+        const rootSideBar = createRewritesSidebar(rewrites, {
+          ...option,
+          path: `${baseDir}${rootDir}`,
+          ignoreList: [...(ignoreList || []), ...localesKeys],
+        });
+        setSideBar(locales["root"].themeConfig, rootSideBar, type);
+      }
     },
   };
 }
 
-const setSideBar = (themeConfig: any, sidebar: DefaultTheme.SidebarMulti) => {
+const setSideBar = (
+  themeConfig: any,
+  sidebar: DefaultTheme.SidebarMulti | DefaultTheme.SidebarItem[],
+  type: SidebarOption["type"]
+) => {
   // 防止 themeConfig 为 undefined
   themeConfig = themeConfig || {};
 
-  themeConfig.sidebar = {
-    ...sidebar,
-    ...(Array.isArray(themeConfig.sidebar) ? logger.warn("自定义 Sidebar 必须是对象形式") : themeConfig.sidebar),
-  };
+  if (type === "object") {
+    themeConfig.sidebar = {
+      ...sidebar,
+      ...(Array.isArray(themeConfig.sidebar) ? logger.warn("自定义 Sidebar 必须是对象形式") : themeConfig.sidebar),
+    };
+  } else {
+    themeConfig.sidebar = [
+      ...(sidebar as DefaultTheme.SidebarItem[]),
+      ...(Object.prototype.toString.call(themeConfig.sidebar) === "[object Object]"
+        ? logger.warn("自定义 Sidebar 必须是数组形式")
+        : themeConfig.sidebar || []),
+    ];
+  }
 
   logger.info("Injected Sidebar Data Successfully. 注入侧边栏数据成功!");
 };